OBC/requirements.txt
numpy>=1.21.0
scikit-learn>=0.24.2
networkx>=2.6.3
torch>=1.9.0
transformers>=4.11.0
PyMuPDF>=1.18.14
python-docx>=0.8.11
pandas>=1.3.0
Pillow>=8.0.0
opencv-python>=4.5.0
nltk>=3.6.0
pytesseract>=0.3.10 